14873 SEC Windows agent's ProgramData directory is accessible only with admins permissions
Previous to this Werk every authenticated Windows user could read some
(uncritical) data from the Windows agent. This was convenient to most
users since one did not need to elevate privileges.

To prevent issues with sensitive data being accidentially written to
logs we restrict the permission to read data of the Windows agent.

To our knowledge it was not possible to exploit this in any way.

We calculated the following CVSS score for this 0.0 (None):
CVSS:3.1/AV:L/AC:L/PR:L/UI:N/S:U/C:N/I:N/A:N

Change-Id: I0a25b8d7c1a675f305f7da1c5298cac8f20791cf