oracle_dataguard_stats: Bufix for wrong Recovery Status on RAC Standby
The check results in false CRITs on RAC standby databases with multiple
running instances.
Some more changes are done, due to wrong design in old code.

The option to monitor the MRP process has been removed, because it is not
needed. The apply_lag is more important then the recovery process itself.
If recovery has been stopped, the monitoring of apply_lag will lead to
ARN/CRIT, which is the most important metric in a Data-Guard environment.

BREAKING Change:
The option mrp state has been removed, because it is not needed. A missing
recovery process (MRP0) is not an issue, until the apply_lag has a proper WARN/CRIT
definition - that is more important then a stopped recovery in the
standby database.

Change-Id: I6049ebb89a19e0a5c717b1fcd0217592740865f4