Add some run time test to ensure we have host attributes registered
Makes it easier to understand import order issues. Without this the
unit test may show some symptoms which are hard to understand:

```
webtest.app.AppError: Bad response: 500 INTERNAL SERVER ERROR (not 200)
'{"title": "Mismatch between endpoint and internal data format. ", "status": 500, "detail": "This could be due to invalid or outdated configuration, or be an error of the implementation. Please check your *.mk files in case you have modified them by hand and run cmk-update-config. If the problem persists afterwards, please report a bug.", "ext": {"errors": {"meta_data": ["Not a string, but a dict", "Tag group name must start with \'tag_\'", "Unknown field."]}, (...)
```

With the change we now get a more meaningful error message which points
us directly to the specific issue.

Of course there are specific unit tests for that, but they fail together
with more than 20 other tests which makes it hard to find the right
track.

Change-Id: I3e90a2fce347936ca75aed4c2cf4b452598f9e34