Extract ApiError to dedicated module
This helps to reduce the scope of openapi related imports triggered
by cmk.gui.context at least a little bit. Now the Endpoint import
does not pull the whole response_schema module anymore.

This is still not ideal, but a small improvement which enables us
to clean up some import cycles later on.

Change-Id: I8e146ec388f16a99e9caa92115fbbaace8eeed8d