14389 Add cmk-passwd utility
This Werk adds the tool cmk-passwd. cmk-passwd can add and change
Checkmk user passwords via the commandline, as could previously be
achieved using htpasswd. Users are advised to use cmk-passwd instead of
htpasswd to manage Checkmk user passwords from now on.

To change a password with cmk-passwd, for example the password of
cmkadmin, simply run
  cmk-passwd cmkadmin
The tool will then prompt the password and ask you to re-type it for
verification. See cmk-passwd -h for further options.

cmk-passwd selects the correct password file location and password
hashing algorithm for the installed version of Checkmk. Behind the
scenes it currently still writes to the same file (etc/htpasswd) in
htpasswd-compatible format. This might be changed in the future though.

Change-Id: If3cb2031566479ed05a7eb827598a6a96f304170