Remove unintelligible tests
They seem to test some of the modes, which would
belong in some component/integration tests.

They may also check that globals are set properly,
which is more or less testing Python and not that
much our own code.

In any case, the amount of monkeypatching and
the assertions in the fixture make these patch
extremely brittle and unintelligible, we are
better off without them. 🚮

Change-Id: I914af99288415e9b86f6ef98c6679fbd3cc38744