Untangle OpenAPI endpoint from UI framework code
The new approach decouples the integration by using hooks. The
disadvantage is that this is more loosely coupled, which makes the
connection invisible for static tooling, but it seems to be OK in
this case.

The permission recording state is now moved to the dedicated module
cmk.gui.plugins.openapi.permission_tracking.

And the actual tracking is implemented in the endpoint. From the body
of the function you can also see that it is tightly coupled with the
endpoint, so we now seem to have the best place.

Change-Id: Ie2279e80ad9dfc1e73020058c35a3f89cf336ccd