14937 Azure: allow selection of services to monitor
This commit adds the possibility to specify which Azure services to
monitor.

You can select the services to monitor in the 'Microsoft Azure'
special agent configuration page.

In case you want to monitor the 'Users in the Active Directory' or the
'AD Connect Sync' services, you will need to grant the
'Directory.Read.All' graph permission to the Azure app and to grant
admin consent to it.

Change-Id: Iaa0e82f8c4d63e6d1f0594a4a28ba9c947f973f4