Fix composition tests: create a site for each module
The composition tests were creating a site with the same id for each
module and this resulted in a crash when saving the results of the site
of the second module because a folder with the same name (site_id) was
already there.
This commit fixes the issue by setting a different site id for each
test module.

Change-Id: I21eddf6f1e1812ab22ca04cf8146f13f4ce87121