14300 Make appliance snapin compatible with Checkmk Appliance firmware 1.6+
When Checkmk is used on on the Checkmk Appliance 1.6 or newer, it is necessary
to update the appliance sidebar snapin.

Previous Checkmk versions displayed the error message 'Failed to render navigation:
Traceback (most recent call last): File: "[stdin]", line 2, in <module>
IOError: [Errno 2] No such file or directory' in the sidebar snapin.

Background: The Checkmk Appliance firmware 1.6 migrates the internal software
stack to Python 3.7 which makes it incompatible with the previous snapin. This
change makes it possible to use Checkmk on older and newer appliance firmware.

Change-Id: Iee4b423091b387e914547523d6fad496710e4b5d