Cleanup view edit transform
It's now clear that render_view_config is called with the actual
configuration of the view or view dashlet and not the value already
prepared for the valuespec editing. The conversion is now handled
internally in that function.

The conversion of the report view element has been separated now since
it's quite a different data structure.

With this change we can remove the load_handler from page_edit_visual
which was only used for converting the view data structure before and
also clean up typing in the view dashlet.

Change-Id: I4046deedbdf7867490de575fcbf2a5eb5628d1aa