Agent controller: Fix privileges drop
Until very recently, the agent controller did not correctly drop
privileges when started eg. as root. To fix this, we started to set the
supplementary group ids, which results in privileges being dropped
correctly.

However, setting the supplementary group ids fails due to insufficient
permissions if we already are the right user (eg. launched by systemd
unit). With this change, we check if we already are the right user and
only attempt to drop privileges if we are not.

Change-Id: I340fab92781971e98c01d1ecee34e0551c2bbc76