Parameterized sorters: Move URL special case to URL parsing
When a user clicks on a column header to change the configured sorting,
the parameterized sorters (currently known as derived column sorters)
need to be handled in a special way. This is now moved to the other URL
parameter parsing code. It now happens earlier at a place you would
expect and reduces the variances during execution.

Change-Id: I6135db78dd0d073dda43220d9f6f2b5822a9131e