14565 FIX inv_if: Fix crash on innovaphone IP811
Previously, the inventory plugin inv_if would crash while
inventorizing interfaces on a innovaphone IP811 with the message
`Parsing of section inv_if failed`. This could potentially happen on
other devices, if they report 'NULL' to indicate missing values.  With
this werk, interfaces with this kind of missing information are
omitted.

SUP-12228

Change-Id: I4e33cb680878852ffb7a341f7374549f96d19286