Remove indirections over HostConfig #1
HostConfig mostly binds a HostName to the ConfigCache.  However,
the callers always already have the hostname available.  Moreover,
this is not used consistently.  This then generates situations
where function get HostConfig, ConfigCache and HostName.

The indirection is therefore useless, obfuscates the call graph,
and is responsible for overly complex function signatures.

CMK-10515

Change-Id: If9b071a424d7c9daa739349cae5b1db2913d1630