Rework DerivedColumnsSorter to ParameterizedSorter
With this change the odd derived columns sorter which was not working
as intended in some cases has been cleaned cleaned up.

* The "Host custom attribute" could only be used as configured sorter in
  the view. Sorting by clicking on a "Host custom attribute" column did
  not work as intended.

* Vice versa for Metric history and forecast sorters. They could be
  configured in the view editor as a dedicated sorter, but it was not
  clear by which column they would sort. Actually they always selected
  the first configured column of that type and sorted by that one.
  Now it is an explicit parameter of that sorter to select one of the
  previously configured columns. The relation is now configured
  explicitly which makes it easier to understand.

  In case such a referenced column is removed while the sorter is kept,
  this is now also validated.

  Interactive sorting by a column while displaying a view already worked
  before.

* The name was related to the "derived columns" feature which is used
  by the "Metric history" painter and "forecast sorters" while it is
  a more generic feature: Enabling sorters to have parameters.

The feature has now been renamed to "Parameterized sorters". This
concept is similar to the parameters which painters can have. In case
of the "Metric history" and "Metric forecast" the painter and sorter
parameters are the same.

The parameters are only used by the cmp method of the Sorters and are
individual for each specific sorter. So the information is now handed
over to the cmp method instead of being stored as a member attribute
which makes the data flow easier to understand.

This change removes many explicit references to the concrete
svc_metrics_hist, svc_metrics_forecast and host_custom_variable sorters
in the generic code.

Along the way it decouples these sorters from the dependency on Cell,
which enables us to clean up the module hierarchy a bit better.

We could do more here to improve the situation even further:

* Cleanup the sorter attribute of SorterSpec to only contain the sorter
  name and separate the parameters into a dedicated attribute
* Move computation of URL encoded sorter references from the generic
  code to the ParameterizedSorter classes.

Change-Id: Id641b22e7539b0168981be795274cbb84fda06df