Cleanup dependency between Cell and sorters
The sorters are only needed for rendering the sort URL to be used in the
header of a column. To have the computation in Cell required us to make
the view spec and sorter implementation available to Cell. Now the sort
URL is computed by the call sites which now simplifies the relation
between Cell and the sorters.

This also improves testability of Cell.

Change-Id: Ibbaf5ba5c533fcde99d191ef336243e1ad69d6dd