14567 FIX KUBE agent_kube: Handle custom Pod condition on GKE
Previously, the Kubernetes special agent would exit with the following
message:

`[special_kube] Agent exited with code 1: Can not convert to timestamp: 'None' of type <class 'NoneType'>`

This would occur if there was a custom Pod condition present in the
cluster with type `cloud.google.com/load-balancer-neg-ready`. With
this werk, the error no longer occurs.

SUP-12170

Change-Id: I444c0c92d478bf2ce534d1a9239f3109cb92718c