Agent receiver: Overwrite verified-uuid header field
The agent receiver relies on the HTTP header field verified-uuid to
verify that the UUID in the URL used to call the endpoint matches the
UUID in the client certificate. This mechanism could be overridden by
manually adding verified-uuid to the HTTP header. As of this commit, we
explicitly override this header field.

Change-Id: I60ffc538b83a7028b99ea447cc015a81bcd68362