introduce testing checks based on snmp walks
Before this change snmp checks were based on string tables. This means
the fetch entry of `register.snmp_section` was never part of the unit
tests. With this change (see test_cisco_wlc_clients.py for an example)
it is possible to test more aspects of `register.snmp_section`.

The idea is that one baseline test for each snmp check can be created by
simply copy and pasting relevant lines from a snmp walk of a real
device. This way we can make sure that the devices we supported once
will be supported in the future.

CMK-8205

Change-Id: I475210fe3d545e93e681c0c419dee0b93493de71