14175 FIX Setup/Move folder: Fixed unknown folder exception after moving folder
Moving a folder always lead to an exception stating the location of the target folder was not found.
The actual move operation always worked.
However, this exception prevented an entry in the audit log and the <tt>Activate changes</tt> page showed no changes to be activated on any site.

Change-Id: I0263ef177f09a25f1ef8c7f37ce33bbdf64bc44e