14921 FIX Add path truststore-path to environment
OpenSSL uses the openssldir option at compile-time for the default
config paths.
Since every site has its own configuration the openssldir cannot be set
to the correct location.

In order to have use the same truststore as the site when using openssl
from the commandline the <tt>SSL_CERT_FILE</tt> variable is now set to
<tt>var/ssl/ca-certificates.crt</tt> for the site user.
So if not specified otherwise the <tt>openssl</tt> command, run as site
user, uses the same truststore as Checkmk.

The Werk is meant for this commit:
3aa300bafacddfd6032e518168eabc59e444b6f3

FEED-7472

Change-Id: I7761158b892845fb94df98bc4adeab5f66c1a877