Create REST API Client
This change is the beginning of a client for the REST API
that makes most of the tests a lot easier to write
by making assumptions that hold true for almost
all tests.

If a certain test does not fit those assumptions,
you can and should use the old API as before.

Change-Id: Ib7392870dd050c01e80ce9d61e5c426b34d9c03e