15031 FIX host_config & dcd: change SiteField field to use configured_sites rather than enabled_sites
Prior to this werk, the REST API endpoint show all hosts returned a 500 error when one of the returned
hosts was part of a disabled remote site. This error also affected the DCD as the connector uses the
same endpoint to retrieve the hosts. This werk resolves this error by changing the site verification
from using all configured sites rather than all enabled sites.

Change-Id: Ia9dff546e17b2a2753f7cb31e3e6001b323ee528