Simplify checkmk packaging
This commit removes the intermediate check_mk-[version].tar.gz which we
used as "Checkmk upstream" package in the early days. Over time the
boundaries between the upstream and OMD packaging were softened,
partially intentionally and partially accidentally. Now it seems to be
the best to break with that concept and complexity to first centralize
everything in the check_mk.make OMD Makefile. From there it's easier to
do the next cleanups.

Change-Id: I199593b66954e6d0d104065a8a4578407d788890