Fix REST API endpoints for modifying site connections (master only)
When creating a new site connection or updating an existinge one, the
REST API did not add the corresponding pending changes. As a result,
after for example adding a new remote site with enabled configuration
replication, there was no pending change to be activated, st. no
configuration synchronization was triggered. Affected endpoints:

/domain-types/site_connection/collections/all
/objects/site_connection/{site_id}

CMK-11635

Change-Id: I6412c7567e6126317304fac06cbdab807a92b76c