15046 FIX elasticsearch_cluster_health: Check plugins used to crash if there were no information about shards and tasks
elasticsearch_cluster_health check plugin for tasks and shards used to crash if there were no information available for the two.
This has now been fixed. If the information is not available, the state of these services turn to UNKNOWN and provides a description that information abput it was not found.

Change-Id: I42155bf1d9574c605f891d56d048ed68d15f00a1