13967 FIX "Topic" attribute for host tags is now optional in REST API
When creating a host tag group using the REST API, it is now optional to specify a topic.
When reading a host tag group without a specifically set topic, it will be displayed as "Tags".
This matches the behaviour of the GUI.

Change-Id: Ifc04cdfa57bd195f6539c423ebb90450da3946cb
(cherry picked from commit 3ef0bfcfbb32de5fcdcc943cf070e0e2dadd000c)