14635 FIX postgres_stat_database: Check plugin used to crash if database size was an empty string
The check plugin used to crash if the database size was an empty string.
This has now been fixed.
From now on if the database size is an empty string, the check state will turn to WARN and in the summary it will say that the database size is not available.

Change-Id: I76526e5cfbf1b7786acbe780de857eed0d81cd4f