13968 FIX Unknown sites in host_config endpoints will now be marked as such
Having a host configured with a unknown site would return a 500 status code error on endpoints such as "/domain-types/host_config/collections/all".
This is common in a distributed monitoring environment.
Instead the site field under the "extensions" and "attributes" key will now return the site id prefixed with "Unknown Site:".
This matched the behaviour of the GUI.

Change-Id: Ib4feae68abd96ab29fc06b1ba0636bfee24b49c9