15033 FIX check_mk_agent.linux: increase refresh interval for chrony section
Prior to this werk, the chrony cached section had the same value for both age validity
and the refresh interval. This led to an occasional CRIT status on the associating service
as this was influenced by the timing of the refresh mechanism, the cache file writeout duration
as well as the section processing duration on Checkmk's side. This werk, therefore, increases
the age validity of the cached section.

Change-Id: I6fac124b883e6632ba63a58ded41ab1d6e57206a