15047 graylog_sources: Added support for a specific timeframe for the messages to be checked in
Added support for a new argument to be set when configuring the graylog special agent.
The argument is "source_since" and allows to set a specific timeframe in which the total number of messages should be checked.
If the argument is not set, the behaviour of the check will stay the same as previosly.
If the argument is set, instead of providing the information about the number of new messages in a specific timeframe, now the total number of messages in a specific timeframe will be provided.

Change-Id: I61f377feb7cb6cd43d3d5af4c7b1fa60b46758dd