13969 FIX ETag for root folder is not correctly generated
The ETag header for the root folder has been generated incorrectly, making it impossible to edit it through the REST API.
This is no longer the case.

Dev note:
The ETag was generated by the werkzeug framework before,
even though it seems to me that no ETag should have been
generated.
As the only way to prevent werkzeug from setting the ETag header
is setting it yourself, I decided to remove the root folder
special case.

SUP-12290

Change-Id: Ifc1895da467a641dc5d0239e097199764ee3c82a