15071 UI: Introduce new toggle icon
Page menu entries which toggle certain view parameters, e.g. "Show checkboxes" in the "All hosts" view, used to show a checkbox icon, that was either empty or checked.
For a more intuitive, consistent and visually understandable user experience, a new toggle icon is introduced and the text of such entries is toggled accordingly.

Change-Id: I2d4ff791221c37d825e3643d6b06c2fb62d14560