15048 Added check plugin for graylog_alerts
Added check plugin for Graylog alerts.
The check plugin will provide information about the Total number of alerts. The number of alerts is limited to 300.
In addition to the above, when configuring the graylog special agent it's possible to set a timeframe to monitor the number of alerts within that timeframe. The timeframe can be set with the "Time for coverage of alerts" option.

Change-Id: Ief34c16497d7f38a599c24ed9da708e089e0494c