Create REST API Client
This change is the beginning of a client for the REST API
that makes most of the tests a lot easier to write
by making assumptions that hold true for almost
all tests.

If a certain test does not fit those assumptions,
you can and should use the old API as before.

Change-Id: I5dc8913431a86c4380ddbfb2a94c0d4cfe0f1288