15043 Setup search: Limit number of results per topic to 80
This is consistent with the Quicksearch default settings and might speed
up the Setup search in certain cases. In case more results are
available, a corresponding message is shown to the user.

Note that the Monitor search already has an implicit limit, namely the
limit of the Quicksearch (configurable via global settings). However,
this limit works fundamentally differently (it limits the overall number
of results across all topics). Therefore, in the current state, we do
not know if there are more results available for a given topic. Hence,
we currently cannot display any message concerning more results after
pressing `Show all results` in the Monitor menu.

CMK-11765

Change-Id: I7e44deb4ddd3d0f4ac225c2d0cc1d09e81d74730