Setup search: Separate search in Redis from permissions check
Searching in Redis is fast, but checking if the current user is allowed
to see the individual results can be slow (might involve rendering a
full page). Therefore, we separate these two steps and do the
permissions check in a generator function. This way, the code rendering
the results can request as many results as it wants to display only,
thus avoiding the permissions check for all results.

CMK-11765

Change-Id: I87a5643c47583d7b5fe30a14c963496c6199c7ab