15041 FIX Improve Setup search performance
WatoMode: Method for checking potentially request-specific permissions

The default implementation only checks static permissions, but
subclasses can add request-specific checks if needed. This method is now
used in the permission checks in the Setup search. This avoids rendering
pages to check if the current user is allowed to see them.

CMK-11763
SUP-12357

Change-Id: I74ab3cdf744643859119afc6877058267e623b55