15063 FIX Resetting password expiry, change enforcement for automation users
When changing the password for an automation user the time of the last password change was not set.
If the global setting <i>Maximum age of passwords</i> was enabled and the password retired, the login was not possible even if the password was newly set.

When a automation user is forced to change the password (e.g. through Werk #14391), being set by an administrator was not honored.

This is now fixed.

Change-Id: I96f6ab8aca447b52f5f5a57eeb4bd9b0acd0c3a9