Revert "Revert "15096 Introducing the Checkmk Development Server""
This reverts commit b288a87c7d1802cae2f36220443ea9dc3541a072.

This is the second attemt and incorporates the following fixes:

Flask session: prevent logout on password change
Flask session: fix flash messages

Original commit message:

15096 Introducing the Checkmk Development Server

This change introduces the GUI development server
which can be run from the command line.

You will need to be in the root of the git repo for
this to work.

Use the following command to get available options
of the server.

$ python -m cmk.gui.dev_server -h

Once it is running, you can log into Checkmk's GUI and
click around. Look at the logs, as they will tell you
where to create your .htpasswd file for your local user(s).

Not supported yet:

 * actually doing monitoring
 * getting status data from CORE
 * sending livestatus commands

For this change to work, the session management of Checkmk
had to be changed to use the one provided by Flask.

Change-Id: I65e85a61eb42844179054ddbfe82355773fb540c