15064 Agent updater checks for certificate validity
The agent updater now reports about the status of the trusted certificates for agent signatures.
The status of those certificates is then checked by the service <i>Check_MK Agent</i> as follows:

LI: Warn if a certificate is corrupt
LI: Warn if a certificate is not valid anymore
LI: Warn if a certificate is about to become invalid
LI: Crit if there is no trusted certificate
LI: Warn/Crit if there will be no valid cert in 90/30 days.

Change-Id: Iece90fe4acc6f734915304e4d0ddc80dd21505e4