Fix serialization of LazyStrings in Visuals
Builtin visuals contain LazyStrings. Copying a builtin visual to the
users' visuals resulted in serializing the LazyString via repr to
l"some string", which cannot be deserialized using literal_eval.
This led to unavailability of the reports page for all users if a user
attempted to modify the default template with their visuals.

LazyStrings are now explictly evaluated to fixed str's when saving
Visuals.

CMK-11706

Change-Id: I3ccb11cd006be260c10c2eac4fc6f2dfc3aa78b4