14688 FIX Increased metric queue sizes
The internal queues of the CMC for communication with the RRD cache daemon,
Carbon connections, and InfluxDB connections has been increased.  This can
help to reduce the "queue full, didn't push update" messages in the CMC log
and the related metric data loss.  Note that this will only help when there
are load peaks or peaks in the number of produced metrics from time to time.
If the monitoring server is not powerful enough to handle the average number
of metrics per second in the long run, there will still be data loss.

Change-Id: I7826433b088d3a2a2f7892b3a172cd691a462c5e