SAML user connector: introduce locked attributes
Some user attributes are managed by the connector (SAML/LDAP/AD) and
should not be editable via the GUI. These attributes should be included
in a list of locked attributes.

"password" is a user attribute that should always be included, since
login is delegated to an identity provider. Other managed attributes are
not yet implemented.

CMK-11931

Change-Id: I812358666047abef2ca7389a8d46ba1ef483115c