REST-API: graph: combine two endpoints
Combine "Get a graph for a metric" and "Get a named Graph" into a single
endpoint.
"Get a named Graph" was only available in cee, but should also be
available in cre.

Also unpack the values of spec directly into the schema. The old schema
was following the internal data structures, but this is not necessary.

Change-Id: Id23e5ee4fd095871ae7e8aa368ece8ef5f1b4aea