15084 kube_replicas: addition of maxUnavailable field for Statefulset with RollingUpdate
The maxUnavailable field for RollingUpdate in Statefulsets is introduced with Kubernetes v1.24.
This werk handles this field and displays the set value in the kube_replicas check. Further
information on this field and how to enable it on Kubernetes' side can be found in the following
article: https://kubernetes.io/blog/2022/05/27/maxunavailable-for-statefulset/

Change-Id: I01b36818bf3b52c6251e9bc830d177e983122621