Bumped googletest to latest version. Create already patched tar file.
Patching sources in CMake seems to be a mess, see e.g. the discussion at
https://gitlab.kitware.com/cmake/cmake/-/issues/17287.  The underlying
problem is: If the unpatched sources and/or the set of patches change,
things should be re-patched from scratch, but only then.  This common
problems seems to be ignore by CMake, which is a pity.  Bazel reportedly has
similar problems, although this sounds strange: The patched sources are a
function of the sources + patches (see above), so Bazel should handle this
just fine.

Anyway: Battling the build system seems to be worse than having this
intermediate step.  If we ever learn how to do this in a cleaner way
*within* the build system, we can easily change our minds again later.

Change-Id: I86c4fe6574c0fde77d177aac71f5da76de8d5f67