14393 Warn about User IDs that become incompatible in 2.2.0
Beginning with Checkmk version 2.2.0, the use of certain special
characters in user IDs will no longer be allowed. As a result, some user
IDs that were previously valid in Checkmk version 2.1.0 will become
invalid in version 2.2.0. This will affect both local users and users
from external connections such as LDAP.

With this Werk you will be warned about users that will become
incompatible in the future. When you run the `omd update` command, it
will display a warning message including the list of affected users.
Moreover, the Setup > Users page in WATO it will also display a warning
message if incompatible user IDs are found.

To be compatible with Checkmk version 2.2.0 please ensure that all user
IDs adhere to the following restrictions: The first character of a user
ID must be a digit, an ASCII letter, an underscore (_), or a dollar sign
($). The rest of the user ID may additionally contain dash (-), at (@),
and dot (.). (In addition to ASCII letters, some umlauts continue to be
permitted).

If you require certain characters that would no longer be allowed and
changing user IDs is not feasible in your system, please contact Checkmk
support for assistance.

CMK-11920

Change-Id: I443d4cb2b2591ecf93b5849f807ecfb15636d0b1