15140 fileinfo: Add the negative age tolerance parameter
Previously, the state would be UNKNOWN if a file has a creation time from the future.
A new parameter 'Negative age tolerance' has been added to 'Size and age of single files'
and 'Size, age and count of file groups' rules.
By setting the parameter, it's possible to ignore the negative age of the file.

SUP-12579

Change-Id: Ibbf594e595789dcf8311992f478fc44928bc7239