15168 KUBE: addition of available to DaemonSet and Deployment replicas
This werk adds the available replicas field of the Deployment and DaemonSet to
the <tt>kube_replicas</tt> check plugin. Similar to the ready replicas, the
duration threshold can be configured via the check plugin's rule.

In addition to the addition, this commit also unifies the replicas
controller section schemas, therefore adjusting the StatefulSet

Change-Id: I9664f02d3e91a6b01ec79425d8b0997204c22b1f