15170 FIX DCD: fix REST API authentication problems
Prior to this werk, the DCD raised a REST API authentication error
when the user's automation password was changed after it was started.
In addition, the same error was also raised when the "Individual
automation user credentials" option in the "Connection to the REST API"
global setting was configured. This werk fixes both error sources.

CMK-10397

Change-Id: I2f58772ffbb89ddf1ed04870c77106c7654bbe44