Github Actions: Do not break google.cloud module
The "Github Actions" job removes non-CRE files to create a test environment
similar to the one on Github. However, the cleanup mechanic was messing around
with all files in the workspace, also in the .venv, .git and .mypy_cache
directories which broke e.g. the google.cloud module.

We now exclude all ".*" directories from the cleanup mechanism to fix this.

Change-Id: I48adfc6fa11b12f96eaacf9cebcc1e780d383bb6