Remove unmaintainable copy-paste between Nagios and cmk
The code was hidden in strings and hard to keep in sync with
changes of the callees.  It was not accessible to linters.
Also, it should be expected that the  "command line" entry
point be used from the command line.  Finally, performance is
barely relevant as this code path is only used once upon
generating a new core configuration.

CMK-12002

Change-Id: I462b1e696ff45dc5746bec38a377dfc10ddeea66